requirements.txt
einops==0.8.0
lpips==0.1.4
matplotlib==3.2.2
numpy==1.18.5
omegaconf==2.3.0
opencv_python==4.10.0.82
Pillow==7.2.0
Pillow==10.4.0
PyYAML==6.0.1
scikit_image==0.16.2
torch==2.0.0+cu118
torchvision==0.15.1+cu118
tqdm==4.64.1
